finish recados part
Finalização da conexão de recados com o banco de dados, inicio da pasta pessoas

+++ src/recados/recados.service.ts
@@ -16,21 +16,21 @@ export class RecadosService {
   private recados: Recado[] = [
     {
       id: 1,
-      texto: "Recado de teste",
-        de: "Zeca",
-      para: "Roberto",
+      texto: 'Recado de teste',
+      de: 'Zeca',
+      para: 'Roberto',
       lido: false,
       data: new Date(),
     },
-  ]
+  ];
 
   throwNotFoundError() {
-    throw new NotFoundException('Recado não encontrado') 
+    throw new NotFoundException('Recado não encontrado');
   }
 
   async findAll() {
     const recados = await this.recadoRepository.find();
-    return recados
+    return recados;
   }
 
   async findOne(id: number) {
@@ -38,62 +38,49 @@ export class RecadosService {
     const recado = await this.recadoRepository.findOne({
       where: {
         id,
-      }
-    })
+      },
+    });
 
-    if(recado) return recado;
+    if (recado) return recado;
 
     this.throwNotFoundError();
   }
 
-  create(createRecadoDto: CreateRecadoDto) {
-    this.lastId++;
-    const id = this.lastId;
+  async create(createRecadoDto: CreateRecadoDto) {
     const novoRecado = {
-      id,
       ...createRecadoDto,
       lido: false,
       data: new Date(),
-    }
-    this.recados.push(novoRecado);
+    };
 
-    if(novoRecado) return novoRecado;
+    const recado = await this.recadoRepository.create(novoRecado);
+
+    return this.recadoRepository.save(recado);
   }
 
-  update(id: number, updateRecadoDto: UpdateRecadoDto) {
-    const recadoExistenteIndex = this.recados.findIndex(item => item.id === id)
+  async update(id: number, updateRecadoDto: UpdateRecadoDto) {
+    const partiallUpdateRecadoDto = {
+      lido: updateRecadoDto?.lido,
+      texto: updateRecadoDto?.texto,
+    };
 
-    if(recadoExistenteIndex < 0) {
-      this.throwNotFoundError();
-    }
+    const recado = await this.recadoRepository.preload({
+      id,
+      ...partiallUpdateRecadoDto,
+    });
 
-    if(recadoExistenteIndex >= 0) {
-      const recadoExistente = this.recados[recadoExistenteIndex];
+    if (!recado) return this.throwNotFoundError();
 
-      this.recados[recadoExistenteIndex] = {
-        id,
-        ...recadoExistente,
-        ...updateRecadoDto,
-      }
-    }
-    const recado = this.recados[ recadoExistenteIndex ]
+    await this.recadoRepository.save(recado);
     return recado;
   }
 
   async remove(id: number) {
-    // const recadoExistenteIndex = this.recados.findIndex(item => item.id === id)
-
-    // if (recadoExistenteIndex < 0) {
-    //   this.throwNotFoundError();
-    // }
-
-    // const recado = this.recados[recadoExistenteIndex]
-
-    // this.recados.splice(recadoExistenteIndex, 1)
-
     const recado = await this.recadoRepository.findOneBy({
-      id
-    })
+      id,
+    });
+
+    if (!recado) return this.throwNotFoundError();
 
     return this.recadoRepository.remove(recado);
   }
